Handle secret validation in Module's configuration
Secrets are moved away from the "secrets" attribute in the Module
module, and into the Module's _configuration attribute
To do this we:
* look for secrets inside the new configuration when attempting to set a
new configuration for the Module. If some secrets are required but not
provided (based on the Module's manifest), we raise an error.
* update the Module's configuration with the relevant values (the
secrets) when an Action sets it un running state