Handle properties validation in Module's config
Secrets are moved away from the "secrets" attribute in the Module
module, and into the Module's _configuration attribute at the same level
as other Module's properties.
We also update the Module's configuration with the relevant values (the
secrets) when an Action sets it un running state.

We also validate candidate Module configuration when the "configuration"
setter is called by looking over the Module's required properties in its
manifest. If some required properties are not present in the candidate
configuration, an error is raised.